requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.10
Markdown==3.3.4
Pillow==8.3.1
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
